settings: Open settings in a dialog from any page
Extract the settings page body into a shared SettingsContent component so the
existing screen and a new global dialog render the same UI with no duplication.

- Add a URL-driven settings dialog opened with the `settings=open` query param,
  mounted in the app layout so it works on any page and on a hard load
- Point the sidebar user menu, the premium API key prompt, and the Slack
  connect hint at the dialog; `/settings` still works as a full screen
- Add an emulated Playwright spec covering the URL and nav entry points
